2025 03 gg withdrawal

+++ org.hl7.fhir.publisher.core/src/main/java/org/hl7/fhir/igtools/web/WebSourceProvider.java
@@ -147,8 +147,10 @@ public class WebSourceProvider {
 
   public void finish(String existingFilesBase, List<String> existingFiles) throws IOException {
     System.out.println("Applying changes to website source at "+source);
-    for (String s : existingFiles) {
-      new File(Utilities.path(s, existingFilesBase, s)).delete();
+    if (existingFiles != null) {
+      for (String s : existingFiles) {
+        new File(Utilities.path(s, existingFilesBase, s)).delete();
+      }
     }
     FileUtilities.copyDirectory(destination, source, null);
     System.out.println("  ... done");
